dl/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
array_record==0.5.1
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.9
dm-tree==0.1.8
docstring_parser==0.16
etils==1.11.0
fastapi==0.115.6
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.4
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
promise==2.3
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
setuptools==75.6.0
simple-parsing==0.1.6
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-metadata==1.16.1
termcolor==2.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
zipp==3.21.0
